Create 'ImageUpload' Seravo-Plugin module
This module has the modules previously known as
optimize-on-upload, sanitize-on-upload and svg-support.
